refactor(frontend): comply with eslint-plugin-react-hooks v7 rules
